Resolve Conflict and Add Selic Module

+++ curve_processor.py
@@ -73,7 +73,8 @@ def interpolate_curve(df):
     if df is None or df.empty:
         return df
 
-    # Ensure index is sorted
+    # Ensure index is sorted and integer
+    df.index = df.index.astype(int)
     df = df.sort_index()
 
     min_du = int(df.index.min())
@@ -100,65 +101,6 @@ def interpolate_curve(df):
 
         i_ant = df.loc[du_ant, 'taxas252']
         i_post = df.loc[du_post, 'taxas252']
-
-        # Multiply by 100?
-        # The scraper returns 'taxas252' as float (e.g. 0.05 for 5%).
-        # The formula uses i_anterior/100. So if input is 5 (percent), we divide by 100.
-        # But scraper output is already divided by 100?
-        # Let's check b3_scraper.py:
-        # taxas252.append(float(texto[i+1].text.strip().replace(',','.')) / 100)
-        # So the dataframe contains 0.10 for 10%.
-        # The formula provided by user: (1 + i_anterior/100).
-        # This implies i_anterior is in percent (e.g. 10.0).
-        # Since our dataframe has raw floats (0.10), we can just use (1 + val).
-        # BUT wait.
-        # Formula: ( (1 + i_anterior/100)^(DU_anterior/252) ... )
-        # User explicitly wrote "i_anterior/100".
-        # If my dataframe has 0.10, then "i_anterior" in the user's mind is 10.
-        # So (1 + 0.10) is correct.
-        # However, the formula ends with "* 100".
-        # If I output the result as 10.0, I need to match the format.
-        # The scraper returns 0.10.
-        # If the user wants the result in the dataframe, usually we keep consistency.
-        # The scraper returns decimals.
-        # The formula returns Percent.
-        # Let's look at the formula again:
-        # ... - 1 ) * 100
-        # This definitely returns a percentage (e.g. 10.5).
-        # But the input dataframe `df` from scraper has decimals (e.g. 0.105).
-        # I should probably return decimals to be consistent with the input dataframe format,
-        # OR I should check if the user specifically requested the output in specific format.
-        # "Armazena este dado em um dataframe que possa ser consultado por um outro código."
-        # The formula explicitly has `* 100` at the end.
-        # And expects `i/100` inside.
-        # So if I plug in `val` (0.10) as `i/100`, that works: `(1 + val)`.
-        # At the end, `... - 1` gives me the decimal rate.
-        # The user formula says `... - 1 ) * 100`.
-        # So the formula produces Percentage.
-        # BUT `b3_scraper.py` produces Decimals.
-        # If I mix them, the output dataframe will have some rows as decimals (the original ones) and some as percentages (the interpolated ones) unless I convert everything.
-        # I MUST be consistent.
-        # I will convert the interpolated values back to decimals (divide by 100) to match the scraper's format,
-        # UNLESS the user implies they want the whole dataframe transformed.
-        # The user provided a formula for the CALCULATION.
-        # The formula returns a Percentage.
-        # If I simply apply the formula, I get percentages.
-        # But I need to fill the gaps in `interpolated_df`.
-        # The existing values in `interpolated_df` are Decimals.
-        # So I should remove the `* 100` from the end of the calculation to store it as Decimal,
-        # OR I should multiply the existing values by 100.
-        # Given `b3_scraper` returns decimals, it is safer to keep decimals.
-        # So I will calculate using the formula logic but keep the result as decimal.
-        # Formula adapted for decimal inputs/outputs:
-        # result_decimal = ( ... )^(252/DU) - 1
-
-        # Let's verify this interpretation.
-        # User formula: ( ... - 1 ) * 100.
-        # If I omit * 100, I get the decimal.
-        # Inside: (1 + i_ant/100).
-        # If i_ant is percent (10), i_ant/100 is 0.10.
-        # My dataframe has 0.10. So I use (1 + df_val).
-        # Correct.
 
         val_ant = i_ant # 0.10
         val_post = i_post # 0.12
@@ -199,14 +141,6 @@ def get_b3_curve_interpolated():
 
     print(f"Interpolating curve for {ref_date}...")
     final_df = interpolate_curve(df)
-
-    # Store reference date in the dataframe (optional but useful)
-    # The scraping function returns index=Days, col=taxas252.
-    # Maybe we want to return the date as well?
-    # The user asked: "Armazena este dado em um dataframe que possa ser consultado por um outro código."
-    # The user didn't explicitly ask for the date column, but it's crucial context.
-    # However, `b3_scraper` returns just rates.
-    # I'll add the date as an attribute or column if needed, but for now just the rates as requested.
 
     return final_df
